pyproject.toml
[tool.poetry]
name = "another-headless-blog"
version = "0.1.0"
description = "Another headless blog"
authors = ["Tommy <tomjleo@gmail.com>"]
license = "GPL-3"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.9"
requests = "^2.26.0"
Markdown = "^3.3.4"
django-filter = "^21.1"
django-extensions = "^3.1.5"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.10.0"
django-rest-knox = "^4.1.0"

[tool.poetry.dev-dependencies]
black = {version = "^21.10b0", allow-prereleases = true}
pytest = "^6.2.5"
pytest-django = "^4.4.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
pdbpp = "^0.10.3"
coverage = "^6.2"
pytest-mock = "^3.6.1"
drf-yasg = "^1.20.0"
ipython = "^7.30.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
